+++ srcs/parse.c
@@ -6,7 +6,7 @@
 /*   By: min-kang <minguk.gaang@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/04 19:01:37 by min-kang          #+#    #+#             */
-/*   Updated: 2022/04/10 19:36:22 by min-kang         ###   ########.fr       */
+/*   Updated: 2022/04/11 14:27:43 by min-kang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -141,7 +141,7 @@ t_map	get_map(int fd)
 		error(4);
 	put_info(&map, map_data);
 	printf("map info : width - %d, height - %d\n", map.width, map.height);
-	printf("there is the person (%d, %d)\n", map.pos.x, map.pos.y);
+	printf("there is the person (%d, %d)\n", (int) map.pos.x, (int) map.pos.y);
 	map.map2d = get_boolmap(map_data, map.width, map.height);
 	return (map);
 }
